Crate rusoto_fsx

Amazon FSx is a fully managed service that makes it easy for storage and application administrators to launch and use shared file storage.

If you're using the service, you're probably looking for FsxClient and Fsx.

Structs

Backup
A backup of an Amazon FSx for Windows File Server file system. You can create a new file system from a backup to protect against data loss.

BackupFailureDetails
If backup creation fails, this structure contains the details of that failure.

CreateBackupRequest
The request object for the CreateBackup operation.

CreateBackupResponse
The response object for the CreateBackup operation.

CreateFileSystemFromBackupRequest
The request object for the CreateFileSystemFromBackup operation.

CreateFileSystemFromBackupResponse
The response object for the CreateFileSystemFromBackup operation.

CreateFileSystemLustreConfiguration
The configuration object for Lustre file systems used in the CreateFileSystem operation.

CreateFileSystemRequest
The request object for the CreateFileSystem operation.

CreateFileSystemResponse
The response object for the CreateFileSystem operation.

CreateFileSystemWindowsConfiguration
The configuration object for the Microsoft Windows file system used in CreateFileSystem and CreateFileSystemFromBackup operations.

DataRepositoryConfiguration
The data repository configuration object for Lustre file systems returned in the response of the CreateFileSystem operation.

DeleteBackupRequest
The request object for DeleteBackup operation.

DeleteBackupResponse
The response object for DeleteBackup operation.

DeleteFileSystemRequest
The request object for DeleteFileSystem operation.

DeleteFileSystemResponse
The response object for the DeleteFileSystem operation.

DeleteFileSystemWindowsConfiguration
The configuration object for the Microsoft Windows file system used in the DeleteFileSystem operation.

DeleteFileSystemWindowsResponse
The response object for the Microsoft Windows file system used in the DeleteFileSystem operation.

DescribeBackupsRequest
The request object for DescribeBackups operation.

DescribeBackupsResponse
Response object for DescribeBackups operation.

DescribeFileSystemsRequest
The request object for DescribeFileSystems operation.

DescribeFileSystemsResponse
The response object for DescribeFileSystems operation.

FileSystem
A description of a specific Amazon FSx file system.

FileSystemFailureDetails
Structure providing details of any failures that occur when creating the file system has failed.

Filter
A filter used to restrict the results of describe calls. You can use multiple filters to return results that meet all applied filter requirements.

FsxClient
A client for the Amazon FSx API.

ListTagsForResourceRequest
The request object for ListTagsForResource operation.

ListTagsForResourceResponse
The response object for ListTagsForResource operation.

LustreFileSystemConfiguration
The configuration for the Amazon FSx for Lustre file system.

Tag
Specifies a key-value pair for a resource tag.

TagResourceRequest
The request object for the TagResource operation.

TagResourceResponse
The response object for the TagResource operation.

UntagResourceRequest
The request object for UntagResource action.

UntagResourceResponse
The response object for UntagResource action.

UpdateFileSystemLustreConfiguration
The configuration object for Amazon FSx for Lustre file systems used in the UpdateFileSystem operation.

UpdateFileSystemRequest
The request object for the UpdateFileSystem operation.

UpdateFileSystemResponse
The response object for the UpdateFileSystem operation.

UpdateFileSystemWindowsConfiguration
The configuration object for the Microsoft Windows file system used in the UpdateFileSystem operation.

WindowsFileSystemConfiguration
The configuration for this Microsoft Windows file system.

Enums

CreateBackupError
Errors returned by CreateBackup

CreateFileSystemError
Errors returned by CreateFileSystem

CreateFileSystemFromBackupError
Errors returned by CreateFileSystemFromBackup

DeleteBackupError
Errors returned by DeleteBackup

DeleteFileSystemError
Errors returned by DeleteFileSystem

DescribeBackupsError
Errors returned by DescribeBackups

DescribeFileSystemsError
Errors returned by DescribeFileSystems

ListTagsForResourceError
Errors returned by ListTagsForResource

TagResourceError
Errors returned by TagResource

UntagResourceError
Errors returned by UntagResource

UpdateFileSystemError
Errors returned by UpdateFileSystem

Traits

Fsx
Trait representing the capabilities of the Amazon FSx API. Amazon FSx clients implement this trait.